ignore the code with `use Module::Functions` in it.
The only purpose of Module::Functions is for exporting all symbols,
including the imported ones, even the ones that are imported
implicitly. For example:

    use Time::Piece;
    use Module::Functions;
    our @EXPORT = get_public_functions();

After this, @EXPORT contains `localtime` function from Time::Piece.
We can no longer state that `Time::Piece` is unused in the current
scope -- because it might be used in a remote location.

Ignoring, has always been an easy option.